fix: refactor add space
