+++ src/context/CustomerContext/CustomerContextProvider/index.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import { ICustomerContext } from "../types";
+
+export const CustomerContext = createContext<ICustomerContext>(
+  {} as ICustomerContext
+);
