+++ src/pages/addProspect/index.tsx
@@ -2,9 +2,9 @@ import { useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useMediaQuery } from "@inubekit/inubekit";
 
+import { CustomerContext } from "@context/CustomerContext/CustomerContextProvider";
 import { Consulting } from "@components/modals/Consulting";
 import { prospectId } from "@mocks/add-prospect/edit-prospect/prospectid.mock";
-import { CustomerContext } from "@context/CustomerContext";
 
 import { stepsAddProspect } from "./config/addProspect.config";
 import { FormData } from "./types";
